backend/iot/simulator/requirements.txt
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
Authlib==1.5.2
avro==1.12.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
confluent-kafka==2.10.0
cryptography==45.0.2
fastavro==1.11.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
minio==7.2.15
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
py4j==0.10.9.7
pycparser==2.22
pycryptodome==3.23.0
pyspark==3.5.5
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
